Moved application, projectData, and questionData models to shared folder
