+++ client/src/app/faculty/post_creation/posts.component.spec.ts
@@ -13,14 +13,14 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatNativeDateModule } from '@angular/material/core';
-import { QuestionData } from 'app/_models/projects/questionData';
+import { QuestionData } from 'app/shared/models/questionData';
 import { MatInputModule } from '@angular/material/input';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { provideRouter } from '@angular/router';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { MatInputHarness } from '@angular/material/input/testing';
-import { ProjectData } from 'app/_models/projects/projectData';
+import { ProjectData } from 'app/shared/models/projectData';
 import { of } from 'rxjs';
 import { AuthService } from 'app/auth/auth-service/auth.service';
 import { MatSelectHarness } from '@angular/material/select/testing';
